requirements.txt
mysqlclient
python-dotenv
djangorestframework
djangorestframework-simplejwt
stripe